+++ internal/server/advkill_test.go
@@ -18,6 +18,7 @@ func killRig(t *testing.T) (*hub, *tracked, map[int32]*tracked) {
 	players := map[int32]*tracked{pl.p.eid: pl}
 	h.playersRef = players
 	w := h.worldFor(0)
+	w.ForceLoad(0, 0, 2) // projectiles fly only through loaded chunks
 	for x := -3; x <= 3; x++ {
 		for z := -3; z <= 12; z++ {
 			w.SetBlock(x, 179, z, worldgen.BlockBase("stone"))